refactor(plugins): remove Plugins route and update routing 🔄
