+++ ui/app/src/model/route.ts
@@ -18,18 +18,8 @@ export const ConfigRoute = '/config';
 export const ImportRoute = '/import';
 export const ProjectRoute = '/projects';
 export const ExploreRoute = '/explore';
-export const PluginsRoute = '/plugins';
 
-const paths = [
-  AdminRoute,
-  SignInRoute,
-  SignUpRoute,
-  ConfigRoute,
-  ImportRoute,
-  ProjectRoute,
-  ExploreRoute,
-  PluginsRoute,
-];
+const paths = [AdminRoute, SignInRoute, SignUpRoute, ConfigRoute, ImportRoute, ProjectRoute, ExploreRoute];
 const prefixPaths = [AdminRoute, ProjectRoute];
 
 export function getBasePathName(): string {
