+++ ui/app/src/views/config/PluginDetailsDialog.tsx
