+++ ui/app/src/views/config/PluginsList.tsx
