+++ ui/endpoint.go
@@ -46,7 +46,6 @@ var (
 		"/import",
 		"/config",
 		"/explore",
-		"/plugins",
 	}
 	capturingPluginName = regexp.MustCompile(`/plugins/([a-zA-Z0-9_-]+)/?.*`)
 )
